feat: Move Settings components in their own folder
