+++ src/renderer/components/Settings/DebugSettings/DebugSettings.css
