+++ src/renderer/components/Settings/DebugSettings/DebugSettings.tsx
@@ -8,7 +8,7 @@ import { DatePicker } from '@mui/x-date-pickers/DatePicker';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
 import dayjs, { Dayjs } from 'dayjs';
 import Stack from '@mui/material/Stack';
-import { electronService } from '../../electron.service';
+import { electronService } from '../../../electron.service';
 import './DebugSettings.css';
 
 const { ipcRenderer } = electronService;
