+++ src/renderer/components/Settings/StashSettings/StashSettings.css
