+++ src/renderer/components/Settings/StashSettings/StashSettings.tsx
@@ -6,7 +6,7 @@ import TextField from '@mui/material/TextField';
 import classNames from 'classnames';
 import './StashSettings.css';
 
-import { electronService } from '../../electron.service';
+import { electronService } from '../../../electron.service';
 import ListItemButton from '@mui/material/ListItemButton';
 import Checkbox from '@mui/material/Checkbox';
 import ListItemText from '@mui/material/ListItemText';
@@ -39,9 +39,9 @@ const StashTabRow = ({ stashTab, indentationLevel = 0 }) => {
     [`Stash-Settings__List-Item--${stashTab.type}`]: true,
     [`Stash-Settings__List-Item--level-${indentationLevel}`]: true,
   });
-  let Icon = require('../../assets/img/tabicons/NormalStash.png');
+  let Icon = require('../../../assets/img/tabicons/NormalStash.png');
   try {
-    Icon = require(`../../assets/img/tabicons/${stashTab.type}.png`);
+    Icon = require(`../../../assets/img/tabicons/${stashTab.type}.png`);
   } catch (e) {
     logger.error('Could not find icon for stash tab type', stashTab.type);
   }
