+++ src/renderer/routes/Settings.tsx
@@ -3,12 +3,13 @@ import { useLoaderData } from 'react-router';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import Box from '@mui/material/Box';
-import StashSettings from '../components/StashSettings/StashSettings';
-import MainSettings from '../components/MainSettings/MainSettings';
-import ItemFilterSettings from '../components/ItemFilterSettings/ItemFilterSettings';
-import DebugSettings from '../components/DebugSettings/DebugSettings';
+import StashSettings from '../components/Settings/StashSettings/StashSettings';
+import MainSettings from '../components/Settings/MainSettings/MainSettings';
+import ItemFilterSettings from '../components/Settings/ItemFilterSettings/ItemFilterSettings';
+import DebugSettings from '../components/Settings/DebugSettings/DebugSettings';
 import './Settings.css';
 import { observer } from 'mobx-react-lite';
+import PricingSettings from '../components/Settings/PricingSettings/PricingSettings';
 
 // Fix to allow for directory selection in inputs
 declare module 'react' {
@@ -47,9 +48,10 @@ const Settings = ({ characterStore, stashTabStore, runStore }) => {
       <Box>
         <Tabs value={tabValue} centered aria-label="Settings Tabs" onChange={handleTabChange}>
           <Tab label="Account" {...a11yProps(0)} />
-          <Tab label="Stashes" {...a11yProps(1)} />
-          <Tab label="Item Filter" {...a11yProps(2)} />
-          <Tab label="Debug" {...a11yProps(3)} />
+          <Tab label="Pricing" {...a11yProps(1)} />
+          <Tab label="Stashes" {...a11yProps(2)} />
+          <Tab label="Item Filter" {...a11yProps(3)} />
+          <Tab label="Debug" {...a11yProps(4)} />
           {/* Add new stuff here */}
         </Tabs>
       </Box>
@@ -57,12 +59,15 @@ const Settings = ({ characterStore, stashTabStore, runStore }) => {
         <MainSettings store={characterStore} settings={settings} runStore={runStore} />
       </div>
       <div hidden={tabValue !== 1}>
-        <StashSettings store={stashTabStore} settings={settings} />
+        <PricingSettings settings={settings} />
       </div>
       <div hidden={tabValue !== 2}>
-        <ItemFilterSettings settings={settings} />
+        <StashSettings store={stashTabStore} settings={settings} />
       </div>
       <div hidden={tabValue !== 3}>
+        <ItemFilterSettings settings={settings} />
+      </div>
+      <div hidden={tabValue !== 4}>
         <DebugSettings runStore={runStore} />
       </div>
     </div>
